added way to filter favorited shows by user

+++ client/src/components/AnimeCard/AnimeCard.js
@@ -7,14 +7,19 @@ import "./AnimeCard.scss";
 
 export default function Card() {
 
-    const [cardArr, setCardArr] = useState([])
+    const [cardArr, setCardArr] = useState([]);
+    const [SessionUserID, setSessionUserID] = useState('')
 
     function formatData() {
         API.getAnime().then(res => {
             let arrData = res.data
             console.log(arrData)
             setCardArr(arrData)
-        })
+        });
+
+        API.getUser().then(res => {
+            setSessionUserID(res.data._id)
+        });
     }
 
     React.useEffect(() => {
@@ -30,22 +35,25 @@ export default function Card() {
 
     return (
         <Grid container direction="row" justify="space-evenly" alignItems="center">
-            {cardArr.map((card) => (
-                <Grid key={card.id} item className="aniGrid">
-                    <Paper elevation={3} className="aniPaper">
-                        <h3 id="title"><a href={card.url} target="_blank" rel="noopener noreferrer">{card.title}</a></h3>
-                        <GridListTile className="imageTile">
-                            <img src={card.image} alt="cardImage" id="cardImage" />
-                        </GridListTile>
-                        <Typography variant="h6" paragraph={true} id="syno">
-                            {card.synopsis}
-                        </Typography>
-                        <Button variant="contained" color="secondary" className="button" startIcon={<DeleteIcon />} onClick={() => handleDelete(card._id)}>
-                            Delete 
-                        </Button>
-                    </Paper>
-                </Grid>
-            ))}
+            {/*  eslint-disable-next-line */}
+            {cardArr.map((card) => {
+                if (card.userID === SessionUserID) return (
+                    <Grid key={card.id} item className="aniGrid">
+                        <Paper elevation={3} className="aniPaper">
+                            <h3 id="title"><a href={card.url} target="_blank" rel="noopener noreferrer">{card.title}</a></h3>
+                            <GridListTile className="imageTile">
+                                <img src={card.image} alt="cardImage" id="cardImage" />
+                            </GridListTile>
+                            <Typography variant="h6" paragraph={true} id="syno">
+                                {card.synopsis}
+                            </Typography>
+                            <Button variant="contained" color="secondary" className="button" startIcon={<DeleteIcon />} onClick={() => handleDelete(card._id)}>
+                                Delete
+                                </Button>
+                        </Paper>
+                    </Grid>
+                )
+            })}
         </Grid>
     )
 }
